fix: Resolve compilation errors from merged PRs
- Fix missing Recommendation and LibraryProfile model classes
- Add missing UpdateModel methods to all AI providers
- Fix method signature mismatches in BrainarrOrchestrator
- Fix IProviderHealthMonitor interface compatibility
- Fix AIProviderFactory CreateProvider method calls
- Fix DiscoveryMode enum usage
- Remove duplicate LibraryProfile class definition
- Fix readonly field assignment errors in providers
- Add missing properties to BrainarrSettings

✅ Main plugin now compiles successfully
⚠️  Test project has 12 remaining compilation errors in test files

+++ Brainarr.Plugin/Services/RateLimiterImproved.cs
@@ -115,7 +115,7 @@ namespace Brainarr.Plugin.Services
             // Statistics
             private long _totalRequests;
             private long _throttledRequests;
-            private double _totalWaitTimeMs;
+            private int _totalWaitTimeMs;
 
             public ResourceRateLimiter(int maxRequests, TimeSpan period, ILogger logger)
             {
@@ -167,7 +167,7 @@ namespace Brainarr.Plugin.Services
                     if (wasThrottled)
                     {
                         Interlocked.Increment(ref _throttledRequests);
-                        Interlocked.Add(ref _totalWaitTimeMs, (long)waitTime.TotalMilliseconds);
+                        Interlocked.Add(ref _totalWaitTimeMs, (int)waitTime.TotalMilliseconds);
                         await Task.Delay(waitTime, cancellationToken).ConfigureAwait(false);
                     }
